refactor: 폴더 경로 및 twmerge 코드 수정
