+++ src/components/modal/ConfirmModal.tsx
@@ -1,5 +1,5 @@
 import Image from 'next/image';
-import Button from '../../Button';
+import Button from '../Button';
 
 type ConfirmModalProps = {
   isOpen: boolean;
