+++ src/components/modal/ReviewModal.tsx
@@ -4,8 +4,8 @@ import Image from 'next/image';
 
 import { formatDate } from '@/src/utils/formatDate';
 import { submitReview } from '@/src/services/createReview';
-import StarRating from '../../StarRating';
-import Button from '../../Button';
+import StarRating from '../StarRating';
+import Button from '../Button';
 
 type ReviewModalProps = {
   isOpen: boolean;
@@ -59,12 +59,7 @@ export default function ReviewModal({
   return (
     <div className="fixed inset-0 flex items-center justify-center z-50">
       <div className="absolute inset-0 bg-black opacity-50" onClick={onClose} />
-      <div
-        className={twMerge(
-          'relative flex flex-col justify-center items-center w-[322px] h-[493px] bg-white rounded-[30px] p-6 shadow-[0_4px_24px_0_#9CB4CA33]',
-          'md:w-[385px] md:h-[549px]'
-        )}
-      >
+      <div className="relative flex flex-col justify-center items-center w-[322px] h-[493px] bg-white rounded-[30px] p-6 shadow-[0_4px_24px_0_#9CB4CA33] md:w-[385px] md:h-[549px]">
         <Image
           src="/images/icons/DeleteIcon.svg"
           alt="리뷰 점수"
